pikeos: disable tree loop distribute patterns
GCC 12 enhanced the tree-loop-distribute-patterns optimization to look for
more code patterns that can be replaced with library calls, for example
memset and strlen. This creates a hidden runtime dependency that is
considered undesirable by many of our customers. The recognition of strlen
and rawmemchr like loops is particularly problematic, as these functions may
not be available on certain bare-metal or cert runtimes.

Closes eng/toolchain/cross-testsuite#9