morello: Move CHERI hardware trap logic to an Ada function with weak fallback
The logic to check which kind of CHERI hardware trap has occurred has been
moved from trap_vector.S to an Ada function in System.CPU_Primitives.
A weak default implementation is provided as a fallback for non-tasking
runtimes to avoid undefined references to the __gnat_raise_cap_*
procedures when linking, as those procedures are not present in the
non-tasking runtimes.

Issue: eng/toolchain/bb-runtimes#11